dojogo-api/requirements.txt
azure-functions>=1.0.0
mysql-connector-python>=8.0.0
PyJWT[crypto]==2.8.0
cryptography>=41.0.0
python-dotenv>=0.19.0
azure-storage-blob>=12.19.0